Unit Tests and Conditionals
Added a UnitTests project and created the #if(Debug) directive.  Need
some help from Pete on understanding when to use each, or if both is
overkill.  UnitTests had problems with testing any of the classes that
inherit Window (i.e. StudentProfile).  This would seem to be the most
useful case since we want to test if the user is an admin, that the
appropriate dialogs are locked (notes, etc.)